Try not to block Hazelcast operation thread too long
Per Hazelcast's document, one or more partitions are dedicated to one
operation thread, any operation that takes long time will block this
thread, and following operations will be blocked. Given that a lock is
acquired and not released for a long time, another acquire attempt
with long timeout on the same lock will block the HZ operation thread.

The change here is to use a smaller timeout value and more attempts to
acquire a permit in semaphore.

Signed-off-by: Richard MENG <mengangr@foxmail.com>